Create README.md
# Attendance-system
This project consists of an attendance system, that captures the users face and updates the attendance with time in an excel sheet. The user names must be the file names that are to be fed as the input. This system can be used in various industries for marking the attendance.